more progress on refactor
