+++ .gitignore
@@ -1,6 +1,9 @@
 # Mac OS files
 *.DS_Store
 
+# VS Code
+.vscode/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
