+++ mlsopt/samplers/pipeline.py
@@ -2,7 +2,7 @@ import logging
 import pandas as pd
 
 # Package imports
-from ..base.samplers import BaseSampler
+from ..base import BaseSampler
 
 __all__ = ["PipelineSampler"]
 _LOGGER = logging.getLogger(__name__)
