+++ mlsopt/utils.py
@@ -1,8 +1,3 @@
-__all__ = [
-    "STATUS_FAIL",
-    "STATUS_OK"
-]
-
 # Status indicators for optimizers
 STATUS_FAIL = 'FAIL'
 STATUS_OK   = 'OK'
