+++ setup.py
@@ -55,7 +55,6 @@ INCLUDE_PACKAGE_DATA = True
 EXTRAS_REQUIRE       = {}
 SCRIPTS              = []
 LICENSE              = "MIT"
-MIN_PYTHON_VERSION   = ">=3.5.*"
 ZIP_SAFE             = False
 CLASSIFIERS          = ['Programming Language :: Python :: 3.5'
                         'Programming Language :: Python :: 3.6',
